+++ crates/mqtt/v3/examples/03_basic_publish_framed_msg.rs
@@ -1,4 +1,4 @@
-use blazebee_mqtt_v4::{
+use blazebee_mqtt_v3::{
     config::{CompressionType, Config, FramingConfig, SerializationConfig, SerializationFormat},
     EndpointMetadata, MqttManager, Publisher,
 };